third API done

+++ RestaurantApplication/src/main/java/com/example/restaurantapplication/model/Product.java
@@ -6,9 +6,9 @@ public class Product
 {
     @Getter private int productID;
     @Getter private String productName;
-    @Getter private int productPrice;
+    @Getter private double productPrice;
 
-    public Product(int productID, String productName, int productPrice)
+    public Product(int productID, String productName, double productPrice)
     {
         this.productID = productID;
         this.productName = productName;
